Update BlockClass1.py
The DrawGrid function creates a working grid and adds part of a path on the grid.  I decided not to finish the path in order to combine my grid and path on a text file.  The CreateArray function takes in a test file filled with ones and zeros separated by commas.  This function takes in the text file and creates a 2D array which is determined by new line spaces and commas.  I will further complete this code in my next commit.