Update BlockClass1.py
My grid matches the numbers in my text file.  Any row and column with an index of zero will now spawn and green texture.  Also, any row and column with an index of one will now spawn and red texture (this creates the path).  This will be my final code before I convert my Turtle graphics to the PyGame graphics package instead.  Lines 59 through 70 are not needed and will likely break the code if the aren't deleted or commented out.