Presentable Working Code For Game 17
Here is some working code before I make my code even more complicated with the addition of a drag and drop system.  You can now add towers by left clicking anywhere (on the left side of the human icon) and then right clicking anywhere on the screen (the right click doesn't need to be on the left side of the screen).